move rom loader to the menu
