+++ cmd/nes/common/sdl.go
@@ -14,5 +14,7 @@ type WindowSize struct {
 type ProgramActions int
 const (
     ProgramToggleSound = iota
-    ProgramLoadRom
+    ProgramQuit
+    ProgramPauseEmulator
+    ProgramUnpauseEmulator
 )
